Saving last changes

+++ src/collatzConjecture.java
@@ -16,7 +16,7 @@ public class collatzConjecture {
                 System.out.println("Please enter a positive integer: ");
                 scr.next();
             }
-            startNum = scr.nextInt();
+            startNum = scr.nextLong();
 
             if (startNum < 1) {
                 System.out.println("Please enter a positive integer: ");
@@ -37,3 +37,43 @@ public class collatzConjecture {
 
 
 }
+
+
+//import java.util.Scanner;
+//import java.math.BigInteger;
+//
+//public class collatzConjecture {
+//
+//    public static void main(String[] args) {
+//
+//        Scanner scr = new Scanner(System.in);
+//        BigInteger startNum;
+//
+//        System.out.println("This program will print Collatz Conjecture");
+//        System.out.println("Please enter a positive integer: ");
+//
+//        do {
+//            while (!scr.hasNextBigInteger()) {
+//                System.out.println("Please enter a positive integer: ");
+//                scr.next();
+//            }
+//            startNum = scr.nextBigInteger();
+//
+//            if (startNum.compareTo(BigInteger.ONE) < 0) {
+//                System.out.println("Please enter a positive integer: ");
+//            }
+//
+//        } while (startNum.compareTo(BigInteger.ONE) < 0);
+//
+//        while (true) {
+//            if (startNum.mod(BigInteger.valueOf(2)).equals(BigInteger.ONE)) {
+//                startNum = startNum.multiply(BigInteger.valueOf(3)).add(BigInteger.ONE);
+//                System.out.println(startNum);
+//            } else {
+//                startNum = startNum.divide(BigInteger.valueOf(2));
+//                System.out.println(startNum);
+//            }
+//        }
+//    }
+//
+//}
